extract editor
