+++ rust/jbob-app/src/lib.rs
@@ -5,6 +5,7 @@ mod styles;
 mod sxfmt;
 mod textbuffer;
 pub mod widgets;
+pub mod sexpr_editor;
 
 pub use events::Event;
 pub use styles::Style;
